extract handler from entrypoint
